+++ Build-Module.ps1
@@ -53,7 +53,9 @@ param(
    [string]$configuration = "LibOnly",
 
    [ValidateSet('Diagnostic', 'Detailed', 'Normal', 'Minimal', 'None', 'ErrorsOnly')]
-   [string]$testOutputLevel = "ErrorsOnly"
+   [string]$testOutputLevel = "ErrorsOnly",
+
+   [switch]$ci
 )
 
 function Import-Pester {
@@ -154,9 +156,11 @@ if (-not $skipLibBuild.IsPresent) {
       New-Item -Path $output\bin -ItemType Directory | Out-Null
    }
 
-   $buildOutput = dotnet build --nologo --verbosity quiet --configuration $configuration | Out-String
+   # Once build machines were updated to .NET 5.0 builds started failing.
+   $buildOutput = dotnet restore --no-cache --ignore-failed-sources --force --force-evaluate | Out-String
+   $buildOutput += dotnet build --nologo --configuration $configuration | Out-String
 
-   if (-not ($buildOutput | Select-String -Pattern 'succeeded')) {
+   if (-not ($buildOutput | Select-String -Pattern 'succeeded') -or $ci.IsPresent) {
       Write-Output $buildOutput
    }