+++ Source/Public/Update-VSTeamWorkItem.ps1
@@ -74,7 +74,7 @@ function Update-VSTeamWorkItem {
       # You must call ConvertTo-Json passing in the value and not
       # not using pipeline.
       # https://stackoverflow.com/questions/18662967/convertto-json-an-array-with-a-single-item
-      $json = ConvertTo-Json @($body) -Compress
+      $json = ConvertTo-Json @($body) -Compress -Depth 100
 
       # Call the REST API
       if ($Force -or $pscmdlet.ShouldProcess($Id, "Update-WorkItem")) {
@@ -82,7 +82,7 @@ function Update-VSTeamWorkItem {
             -Area wit `
             -Resource workitems `
             -id $Id `
-            -ContentType 'application/json-patch+json' `
+            -ContentType 'application/json-patch+json; charset=utf-8' `
             -Body $json `
             -Version $(_getApiVersion Core)