+++ Tests/function/tests/Get-VSTeamRelease.Tests.ps1
@@ -79,5 +79,17 @@ Describe 'VSTeamRelease' {
             $Uri -eq "https://vsrm.dev.azure.com/test/_apis/release/releases?api-version=$(_getApiVersion Release)"
          }
       }
+
+      It 'with build Id should return release as Object' {
+         ## Act
+         $r = Get-VSTeamRelease -ProjectName VSTeamRelease -artifactVersionId 101
+
+         ## Assert
+         $r | Get-Member | Select-Object -First 1 -ExpandProperty TypeName | Should -Be 'vsteam_lib.Release'
+
+         Should -Invoke Invoke-RestMethod -Exactly -Scope It -Times 1 -ParameterFilter {
+            $Uri -eq "https://vsrm.dev.azure.com/test/VSTeamRelease/_apis/release/releases?api-version=$(_getApiVersion Release)&artifactVersionId=101"
+         }
+      }
    }
 }