+++ Tests/function/tests/Set-VSTeamPipelineBilling.Tests.ps1
@@ -3,6 +3,8 @@ Set-StrictMode -Version Latest
 Describe 'VSTeamPipelineBilling' -Tag 'unit', 'billing' {
    BeforeAll {
       . "$PSScriptRoot\_testInitialize.ps1" $PSCommandPath
+      . "$baseFolder/Source/Public/Get-VSTeamUserProfile.ps1"
+      . "$baseFolder/Source/Public/Get-VSTeamAccounts.ps1"
 
       # Set the account to use for testing. A normal user would do this
       # using the Set-VSTeamAccount function.
@@ -12,7 +14,13 @@ Describe 'VSTeamPipelineBilling' -Tag 'unit', 'billing' {
          return $null
       }
 
+      $accounts = Open-SampleFile 'Get-VSTeamAccounts.json'
+
       Mock _getBillingToken { Open-SampleFile 'getBillingToken.json' }
+      Mock _getInstance { return "https://dev.azure.com/TestOrg01" }
+      Mock Get-VSTeamUserProfile { Open-SampleFile 'Get-VSTeamUserProfile.json' }
+      Mock Get-VSTeamAccounts { return $accounts }
+      Mock _getApiVersion { return '5.1-preview.1' } -ParameterFilter { $Service -eq 'Billing' }
    }
 
    Context 'Set-VSTeamPipelineBilling' {
@@ -20,19 +28,31 @@ Describe 'VSTeamPipelineBilling' -Tag 'unit', 'billing' {
          Set-VSTeamPipelineBilling `
             -Type "HostedPipeline" `
             -OrganizationId "68c631ce-4886-4825-a471-94a74fb6ecda" `
-            -SubscriptionId "fbafed90-9c59-4889-8187-8e74e2cf06e7" `
             -Quantity 2 `
             -Force
 
          Should -Invoke Invoke-RestMethod -Exactly -Scope It -Times 1 `
             -ParameterFilter {
-            $Method -eq 'Post' -and
-            $Body -like '*"azureSubscriptionId":"fbafed90-9c59-4889-8187-8e74e2cf06e7"*' -and
-            $Body -like '*"committedQuantity":2*' -and
-            $Body -like '*"galleryId":"ms.build-release-hosted-pipelines"*' -and
-            $Uri -like "https://commerceprodwus21.vscommerce.visualstudio.com/_apis/OfferSubscription/OfferSubscription?api-version=5.1-preview.1*" -and
-            $Uri -like "*billingTarget=68c631ce-4886-4825-a471-94a74fb6ecda*" -and
-            $Uri -like "*skipSubscriptionValidation=True*"
+            $Method -eq 'Patch' -and
+            $Body -like '*"purchaseQuantity":2*' -and
+            $Body -like '*"meterId":"4bad9897-8d87-43bb-80be-5e6e8fefa3de"*' -and
+            $Uri -like "https://azdevopscommerce.dev.azure.com/68c631ce-4886-4825-a471-94a74fb6ecda/_apis/AzComm/MeterResource?api-version=5.1-preview.1*"
+         }
+      }
+
+      It 'should set hosted pipeline for default connected Org' {
+         Set-VSTeamPipelineBilling `
+            -Type "HostedPipeline" `
+            -Quantity 2 `
+            -Force
+
+         Should -Invoke Invoke-RestMethod -Exactly -Scope It -Times 1 `
+            -ParameterFilter {
+            $Method -eq 'Patch' -and
+            $Body -like '*"purchaseQuantity":2*' -and
+            $Body -like '*"meterId":"4bad9897-8d87-43bb-80be-5e6e8fefa3de"*' -and
+            $Uri -like "https://azdevopscommerce.dev.azure.com/81d6e09f-266a-4dd2-886c-b3d62341681e/_apis/AzComm/MeterResource?api-version=5.1-preview.1*"
+
          }
       }
 
@@ -40,20 +60,28 @@ Describe 'VSTeamPipelineBilling' -Tag 'unit', 'billing' {
          Set-VSTeamPipelineBilling `
             -Type "PrivatePipeline" `
             -OrganizationId "68c631ce-4886-4825-a471-94a74fb6ecda" `
-            -SubscriptionId "fbafed90-9c59-4889-8187-8e74e2cf06e7" `
             -Quantity 2 `
             -Force
 
          Should -Invoke Invoke-RestMethod -Exactly -Scope It -Times 1 `
             -ParameterFilter {
-            $Method -eq 'Post' -and
-            $Body -like '*"azureSubscriptionId":"fbafed90-9c59-4889-8187-8e74e2cf06e7"*' -and
-            $Body -like '*"committedQuantity":2*' -and
-            $Body -like '*"galleryId":"ms.build-release-private-pipelines"*' -and
-            $Uri -like "https://commerceprodwus21.vscommerce.visualstudio.com/_apis/OfferSubscription/OfferSubscription?api-version=5.1-preview.1*" -and
-            $Uri -like "*billingTarget=68c631ce-4886-4825-a471-94a74fb6ecda*" -and
-            $Uri -like "*skipSubscriptionValidation=True*"
+            $Method -eq 'Patch' -and
+            $Body -like '*"purchaseQuantity":3*' -and
+            $Body -like '*"meterId":"f44a67f2-53ae-4044-bd58-1c8aca386b98"*' -and
+            $Uri -like "https://azdevopscommerce.dev.azure.com/68c631ce-4886-4825-a471-94a74fb6ecda/_apis/AzComm/MeterResource?api-version=5.1-preview.1*"
          }
+      }
+
+      It 'should throw' {
+         Mock _getApiVersion { return '' } -ParameterFilter { $Service -eq 'Billing' }
+
+         {
+            Set-VSTeamPipelineBilling `
+            -Type "PrivatePipeline" `
+            -OrganizationId "68c631ce-4886-4825-a471-94a74fb6ecda" `
+            -Quantity 2 `
+            -Force
+         } | Should -Throw
       }
    }
 }